test: 合并 security 双文件
- test_security_extended.py 30 例中 28 例与 test_security.py 重复，删除
- 2 个独有用例（构造初始化、verify_password 异常路径）并入 test_security.py
